remove Current singleton and replace with static class
